requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings 
python-dotenv 
pymysql
cryptography
alembic
bcrypt
python-jose[cryptography]
email-validator
